feat: editor settings entry
