+++ apps/web/app/components/settings/localization/View.vue
@@ -1,13 +1,6 @@
 <template>
   <SiteConfigurationView>
     <template #setting-title>{{ getEditorTranslation('label') }}</template>
-    <template #setting-description>
-      <div class="flex flex-col px-4 text-sm">
-        <p class="pb-2">
-          <span class="align-middle font-bold">{{ getEditorTranslation('description') }}</span>
-        </p>
-      </div>
-    </template>
   </SiteConfigurationView>
 </template>
 
@@ -16,12 +9,10 @@
 <i18n lang="json">
 {
   "en": {
-    "label": "Localization",
-    "description": "Manage translations for static texts inside the shop."
+    "label": "Localization"
   },
   "de": {
-    "label": "Localization",
-    "description": "Manage translations for static texts inside the shop."
+    "label": "Localization"
   }
 }
 </i18n>
